backend/requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.27.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
pgvector==0.3.6
aiofiles==24.1.0